update user CRUD functionality so it interfaces properly with new db collection schemas

+++ app/api/users/route.ts
@@ -4,7 +4,7 @@ import { dbConnect } from "../../lib/dbConnect";
 import { verifyToken } from "../../lib/authFunctions";
 import { NextResponse } from "next/server";
 
-// gets all water users' data
+// gets all water users' personal info
 export async function POST(req: Request) {
     try {
         const { token, pathname } = await req.json();
@@ -15,7 +15,7 @@ export async function POST(req: Request) {
             return NextResponse.json({ users: [], validity });
         };
 
-        const collection = await waterUsersCollection();
+        const collection = await collectionConnect('waterUsers');
         const cursor = collection?.find({});
         const users = await cursor?.toArray();
         await cursor?.close();
@@ -29,7 +29,7 @@ export async function POST(req: Request) {
 };
 
 
-// edits an existing water user
+// edits an existing water user's contact info
 export async function PATCH(req: Request) {
     try {
         const {
@@ -44,17 +44,20 @@ export async function PATCH(req: Request) {
             return NextResponse.json({ success: false, validity });
         };
 
-        const collection = await waterUsersCollection();
+        const collection = await collectionConnect('waterUsers');
         const cursor = await collection.updateOne(
             { _id: new ObjectId(updateInfo._id) },
-            { $set: { info: updateInfo.info } }
+            { $set: { 
+                'info.name': updateInfo.info.name,
+                'info.email': updateInfo.info.email,
+                'info.address': updateInfo.info.address
+            }}
         );
 
-        if (cursor.modifiedCount == 1){
-            return NextResponse.json({ success: true, validity });
-        } else {
+        if (cursor.modifiedCount == 0) {
             return NextResponse.json({ success: false, validity });
         };
+        return NextResponse.json({ success: true, validity });
 
     } catch (error) {
         console.log(`error thrown in [/api/users] PATCH: ` + error);
@@ -76,21 +79,74 @@ export async function PUT(req: Request) {
             return NextResponse.json({ success: false, validity, newUser: undefined });
         };
 
-        const collection = await waterUsersCollection();
-        const cursor = await collection.insertOne({info: newUserInfo});
-
-        if (!cursor.insertedId) {
+        // insert new water user document into the waterUsers collection
+        let collection = await collectionConnect('waterUsers');
+        let cursor = await collection.insertOne({ info: newUserInfo });
+        const userId = cursor.insertedId;
+        if (!userId) {
+            console.log('failed to insert water user into waterUsers collection');
             return NextResponse.json({ success: false, validity, newUser: undefined });
-        } else {
-            return NextResponse.json({
-                success: true,
-                validity,
-                newUser: {
-                    _id: cursor.insertedId,
-                    info: newUserInfo
-                }
-            });
         };
+
+        // insert water usage data document into usage collection
+        collection = await collectionConnect('usage');
+        cursor = await collection.insertOne({
+            _id: userId,
+            prev: {
+                Q1: { 1: 0, 2: 0, 3: 0 },
+                Q2: { 4: 0, 5: 0, 6: 0 },
+                Q3: { 7: 0, 8: 0, 9: 0 },
+                Q4: { 10: 0, 11: 0, 12: 0 }
+            },
+            cur: {
+                Q1: { 1: 0, 2: 0, 3: 0 },
+                Q2: { 4: 0, 5: 0, 6: 0 },
+                Q3: { 7: 0, 8: 0, 9: 0 },
+                Q4: { 10: 0, 11: 0, 12: 0 }
+            }
+        });
+        if (!cursor.acknowledged) {
+            console.log('failed to insert new water usage data document in usage collection');
+            collection = await collectionConnect('waterUsers');
+            const cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
+            if (cursor.deletedCount == 0) {
+                console.log(`failed to delete inserted water user ${userId} from waterUsers collection. direct db modification required...`)
+            };
+            return NextResponse.json({ success: false, validity, newUser: undefined });
+        }
+
+        // insert new balance history document into balance collection
+        collection = await collectionConnect('balances');
+        cursor = await collection.insertOne({
+            _id: userId,
+            prev: {},
+            cur: {
+                'new user creation': newUserInfo.balance
+            }
+        })
+        if (!cursor.acknowledged) {
+            console.log('failed to insert new water usage balance document in balance collection');
+            collection = await collectionConnect('waterUsers');
+            let cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
+            if (cursor.deletedCount == 0) {
+                console.log(`failed to delete inserted water user ${userId} from waterUsers collection. direct db modification required...`)
+            };
+            collection = await collectionConnect('usage');
+            cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
+            if (cursor.deletedCount == 0) {
+                console.log(`failed to delete inserted water user ${userId} from usage collection. direct db modification required...`)
+            };
+            return NextResponse.json({ success: false, validity, newUser: undefined });
+        };
+
+        return NextResponse.json({
+            success: true,
+            validity,
+            newUser: {
+                _id: userId,
+                info: newUserInfo
+            }
+        });
 
     } catch (error) {
         console.log(`error thrown in [/api/users] PUT: ` + error);
@@ -99,7 +155,7 @@ export async function PUT(req: Request) {
 };
 
 
-// deleted a water user's info (but not the user's historical water usage data)
+// deletes a water user
 export async function DELETE(req: Request) {
     try {
         const { token, pathname, userId }: {
@@ -112,14 +168,28 @@ export async function DELETE(req: Request) {
             return NextResponse.json({ success: false, validity });
         };
 
-        const collection = await waterUsersCollection();
-        const cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
-
+        let collection = await collectionConnect('waterUsers');
+        let cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
         if (cursor.deletedCount == 0) {
+            console.log('failed to delete user from waterUsers collection. direct db modification required...');
             return NextResponse.json({ success: false, validity });
-        } else {
-            return NextResponse.json({ success: true, validity });
-        };
+        }
+
+        collection = await collectionConnect('usage');
+        cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
+        if (cursor.deletedCount == 0) {
+            console.log('failed to delete user from waterUsers collection. direct db modification required...');
+            return NextResponse.json({ success: false, validity });
+        }
+
+        collection = await collectionConnect('balances');
+        cursor = await collection.deleteOne({ _id: new ObjectId(userId) });
+        if (cursor.deletedCount == 0) {
+            console.log('failed to delete user from waterUsers collection. direct db modification required...');
+            return NextResponse.json({ success: false, validity });
+        }
+
+        return NextResponse.json({ success: true, validity });
 
     } catch (error) {
         console.log(`error thrown in [/api/users] PUT: ` + error);
@@ -155,9 +225,9 @@ const validateRequest = async (token: string, pathname: string, method: string) 
 /**
  * @returns the waterUsers collection in the waterUsersDb database.
  */
-const waterUsersCollection = async () => {
+const collectionConnect = async (name: string) => {
     const dbClient = await dbConnect();
     const db = dbClient?.db('waterUsersDb');
-    const collection = db?.collection('waterUsers');
+    const collection = db?.collection(name);
     return collection;
 };
